Move subject for a new file and create aux folder, with window-factory (updated spotify-player with these new files)
